Move plantation evaluations to criteria
