+++ envergo/templates/haie/moulinette/plantation_evaluation/inadequate.html
@@ -2,6 +2,7 @@
   <h3 class="fr-alert__title">La plantation envisagée est insuffisante</h3>
   <p>Le projet de plantation décrit n'apparaît pas acceptable au regard des haies détruites :</p>
   <ul>
+    {% for condition in plantation_evaluation.invalid_conditions %}<li>{{ condition.text }}</li>{% endfor %}
     {% if "length_to_plant" in plantation_evaluation.unfulfilled_conditions %}
       <li>
         La longueur de la haie à planter est insuffisante. Elle doit être supérieure à
